requirements.txt
python-decouple
discord.py
humanfriendly
pytz